refactor(components): extract action components into an action layout and action
